Successfully understand file system of unix like system
The simplest idea that everything in unix like system is file means
terminal is also a file. If I want to print something on the terminal,
I must print to the terminal file(/dev/tty). If someone close it, just
reopen it with freopen("/dev/tty","a+b",stdout); will be fine.

And I added enough comments in the source code.
I removed some debug information that I don’t need.